+++ packages/hooks/package.json
@@ -30,6 +30,7 @@
     "@lightdotso/params": "workspace:*",
     "@lightdotso/query": "workspace:*",
     "@lightdotso/schemas": "workspace:*",
+    "@lightdotso/sdk": "workspace:*",
     "@lightdotso/sequence": "workspace:*",
     "@lightdotso/stores": "workspace:*",
     "@lightdotso/types": "workspace:*",
